% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMH.R
\name{simpleMH}
\alias{simpleMH}
\title{Simple Metropolis-Hastings MCMC}
\usage{
simpleMH(f, inits, theta.cov, max.iter, coda = FALSE, ...)
}
\arguments{
\item{f}{function that returns a single scalar value proportional to the log
probability density to sample from.}

\item{inits}{numeric vector with the initial values for the parameters to
estimate}

\item{theta.cov}{covariance matrix of the parameters to estimate.}

\item{max.iter}{maximum number of function evaluations}

\item{coda}{logical. Should the samples be returned as \link[coda:mcmc]{coda::mcmc}
object? (defaults to \code{FALSE})}

\item{...}{further arguments passed to \code{f}}
}
\value{
\itemize{
\item if \code{coda = FALSE} a list with:
\itemize{
\item \emph{samples}: A three dimensional array of samples with dimensions \code{walker}
x \code{generation} x \code{parameter}
\item \emph{log.p}: A matrix with the log density evaluate for each walker at each
generation.
}
\item if \code{coda = TRUE} a list with:
\itemize{
\item \emph{samples}: A object of class \link[coda:mcmc.list]{coda::mcmc.list} containing all samples.
\item \emph{log.p}: A matrix with the log density evaluate for each walker at each
generation.
}
}
}
\description{
Simple Metropolis-Hastings MCMC
}
